# Moved curl_multi_wait call from event-loop to outer loop